Move `Batch` class into `pypers.batch` module
