+++ tests/test_task.py
@@ -1,9 +1,7 @@
 import dill
 import gzip
 import json
-import os
 import pathlib
-import pprint
 import tempfile
 import unittest
 from unittest.mock import (
@@ -964,162 +962,3 @@ class Task__run(unittest.TestCase):
         )
         self.assertEqual(mock_create_pipeline.return_value.process.call_count, 2)
         mock_store.assert_called_once()
-
-
-def create_task_file(task_path: pypers.task.PathLike, spec_yaml: str) -> None:
-    task_path = pathlib.Path(task_path)
-    task_filepath = task_path / 'task.yml'
-    if not task_path.is_dir():
-        task_path.mkdir(parents = True, exist_ok = True)
-    with task_filepath.open('w') as spec_file:
-        spec_file.write(spec_yaml)
-
-
-class Batch__task(unittest.TestCase):
-
-    def test_virtual_paths(self):
-        batch = pypers.task.Batch()
-
-        # Load the tasks
-        task1 = batch.task(
-            path = 'path/to/task',
-            spec = dict(),
-        )
-        task2 = batch.task(
-            path = 'path/to/task/2',
-            spec = dict(),
-        )
-
-        # Verify the tasks
-        self.assertEqual(len(batch.tasks), 2)
-        self.assertIsNone(task1.parent)
-        self.assertIs(task2.parent, task1)
-
-    def test_virtual_path_without_parent(self):
-        batch = pypers.task.Batch()
-        task = batch.task(
-            path = '',
-            spec = dict(),
-        )
-        self.assertIsNone(task.parent)
-
-    @testsuite.with_temporary_paths(1)
-    def test_spec_files(self, root):
-        create_task_file(root, 'pipeline: pypers.pipeline.Pipeline')
-        create_task_file(root / 'task-2', 'field1: value1')
-
-        # Load the tasks
-        batch = pypers.task.Batch()
-        task1 = batch.task(path = root)
-        task2 = batch.task(path = root / 'task-2')
-
-        # Verify the number of loaded tasks
-        self.assertEqual(len(batch.tasks), 2)
-
-        # Verify task1
-        self.assertIsNone(task1.parent)
-        self.assertEqual(task1.full_spec, dict(pipeline = 'pypers.pipeline.Pipeline'))
-
-        # Verify task2
-        self.assertIs(task2.parent, task1)
-        self.assertEqual(task2.full_spec, dict(pipeline = 'pypers.pipeline.Pipeline', field1 = 'value1'))
-
-    @testsuite.with_temporary_paths(1)
-    def test_spec_files_with_override(self, root):
-        create_task_file(root, 'pipeline: pypers.pipeline.Pipeline')
-        batch = pypers.task.Batch()
-        task = batch.task(
-            path = root,
-            spec = dict(pipeline = 'pypers.pipeline.Pipeline2'),
-        )
-        self.assertIsNone(task.parent)
-        self.assertEqual(task.spec, dict(pipeline = 'pypers.pipeline.Pipeline2'))
-
-    @testsuite.with_temporary_paths(1)
-    def test_mixed_virtual_paths_and_spec_files(self, root):
-        create_task_file(root, 'pipeline: pypers.pipeline.Pipeline')
-        create_task_file(root / 'task-2', 'field1: value1')
-
-        # Load the tasks
-        batch = pypers.task.Batch()
-        task3 = batch.task(
-            path = root / 'task-2' / 'task-3',
-            spec = dict(field2 = 'value2'),
-        )
-        task2 = batch.task(path = root / 'task-2')
-
-        # Verify the number of loaded tasks
-        self.assertEqual(len(batch.tasks), 3)
-
-        # Verify task3
-        self.assertIs(task3.parent, task2)
-        self.assertEqual(task3.full_spec, dict(pipeline = 'pypers.pipeline.Pipeline', field1 = 'value1', field2 = 'value2'))
-
-
-class Batch__load(unittest.TestCase):
-
-    def setUp(self):
-        self.tempdir = tempfile.TemporaryDirectory()
-        self.root_path = pathlib.Path(self.tempdir.name)
-        create_task_file(self.root_path, 'pipeline: pypers.pipeline.Pipeline')
-        create_task_file(self.root_path / 'task-2', 'field1: value1')
-        create_task_file(self.root_path / 'task-2' / 'task-3', 'field2: value2')
-        self.batch = pypers.task.Batch()
-
-    def tearDown(self):
-        self.tempdir.cleanup()
-
-    def test(self):
-        self.batch.load(self.root_path)
-        self.assertEqual(
-            [task.path for task in self.batch.tasks.values()],
-            [
-                self.root_path,
-                self.root_path / 'task-2',
-                self.root_path / 'task-2' / 'task-3',
-            ],
-        )
-
-    def test_illegal_path(self):
-        with self.assertRaises(AssertionError):
-            self.batch.load(self.root_path / 'task-3')
-
-
-class Batch__run(unittest.TestCase):
-
-    stage1_cls = testsuite.create_stage_class(id = 'stage1')
-    stage2_cls = testsuite.create_stage_class(id = 'stage2')
-
-    def setUp(self):
-        self.tempdir = tempfile.TemporaryDirectory()
-        self.root_path = pathlib.Path(self.tempdir.name)
-        create_task_file(
-            self.root_path,
-            'runnable: true' '\n'
-            'pipeline:' '\n'
-            '- tests.test_task.Task__create_pipeline.stage1_cls' '\n'
-            '- tests.test_task.Task__create_pipeline.stage2_cls' '\n'
-        )
-        create_task_file(
-            self.root_path / 'task-2',
-            'stage1:' '\n'
-            '  key1: value1' '\n'
-        )
-        create_task_file(
-            self.root_path / 'task-3',
-            'stage2:' '\n'
-            '  key2: value2' '\n'
-        )
-        self.batch = pypers.task.Batch()
-        self.batch.load(self.root_path)
-        self.testsuite_pid = os.getpid()
-
-    def tearDown(self):
-        if os.getpid() == self.testsuite_pid:
-            self.tempdir.cleanup()
-
-    @testsuite.with_temporary_paths(1)
-    def test(self, path):
-        status = pypers.status.Status(path = path)
-        self.batch.run(status = status)
-        self.assertEqual([list(item.keys()) for item in status.data], [['expand']] * 3, '\n' + pprint.pformat(status.data))
